requirements.txt
absl-py==0.12.0
anyio==2.2.0
appdirs==1.4.4
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
Babel==2.9.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.9.1
black==20.8b1
bleach==3.3.0
boto3==1.17.39
botocore==1.20.39
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.0
chardet==3.0.4
click==7.1.2
conda==4.9.2
conda-build==3.18.11
conda-package-handling==1.7.0
cryptography==2.9.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.7.1
efficientnet-pytorch==0.7.0
entrypoints==0.3
filelock==3.0.12
fire==0.4.0
future==0.18.2
gevent==21.1.2
glob2==0.7
google-auth==1.28.0
google-auth-oauthlib==0.4.3
greenlet==1.0.0
grpcio==1.36.1
idna==2.9
importlib-metadata==3.9.0
inotify-simple==1.2.1
ipykernel==5.5.0
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.17.1
Jinja2==2.11.2
jmespath==0.10.0
joblib==1.0.1
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyter-packaging==0.7.12
jupyter-server==1.5.1
jupyterlab==3.0.12
jupyterlab-pygments==0.1.2
jupyterlab-server==2.3.0
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
libarchive-c==2.9
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
mkl-fft==1.1.0
mkl-random==1.1.1
mkl-service==2.3.0
mypy-extensions==0.4.3
nbclassic==0.2.6
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.3.0
numpy==1.18.5
oauthlib==3.1.0
olefile==0.46
opencv-python==4.5.1.48
packaging==20.9
pandas==1.1.5
pandocfilters==1.4.3
paramiko==2.7.2
parso==0.7.0
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pkginfo==1.5.0.1
prometheus-client==0.9.0
prompt-toolkit==3.0.5
protobuf==3.15.6
psutil==5.7.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser==2.20
Pygments==2.6.1
PyNaCl==1.4.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
pyzmq==22.0.3
qtconsole==5.0.3
QtPy==1.9.0
regex==2021.3.17
requests==2.23.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.7.2
ruamel-yaml==0.15.87
s3transfer==0.3.6
sagemaker-training==3.7.3
scikit-learn==0.24.1
scipy==1.6.2
Send2Trash==1.5.0
six==1.14.0
sniffio==1.2.0
soupsieve==2.0.1
tabulate==0.8.9
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
terminado==0.9.3
testpath==0.4.4
threadpoolctl==2.1.0
timm==0.4.5
toml==0.10.2
torch==1.6.0
torchvision==0.7.0
tornado==6.1
tqdm==4.46.0
traitlets==4.3.3
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.25.8
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
zipp==3.4.1
zope.event==4.5.0
zope.interface==5.3.0
